Time: 18 ms (50.39%) | Memory: 64.1 MB (67.90%) - LeetSync

+++ 969-number-of-recent-calls/number-of-recent-calls.cpp
@@ -11,7 +11,7 @@ public:
         while (t - q.front() > 3000) {
             q.pop();
         }
-        return q.size();
+        return static_cast<int>(q.size());
     }
 };
